+++ packages/adp-tooling/test/unit/writer/index.test.ts
@@ -8,7 +8,7 @@ import { rimraf } from 'rimraf';
 describe('ADP writer', () => {
     const fs = create(createStorage());
     const debug = !!process.env['UX_DEBUG'];
-    const outputDir = join(__dirname, '../../fixtures/test');
+    const outputDir = join(__dirname, '../../fixtures/test-output');
 
     beforeAll(async () => {
         await rimraf(outputDir);
@@ -40,6 +40,51 @@ describe('ADP writer', () => {
             const projectDir = join(outputDir, 'minimal');
             await generate(projectDir, config, fs);
             expect(fs.dump(projectDir)).toMatchSnapshot();
+        });
+
+        test('add deploy config', async () => {
+            const projectDir = join(outputDir, 'deploy');
+            await generate(
+                projectDir,
+                {
+                    ...config,
+                    deploy: {
+                        package: '$TMP'
+                    }
+                },
+                fs
+            );
+            expect(
+                fs.dump(
+                    projectDir,
+                    (file) => file.dirname === projectDir && ['package.json', 'ui5-deploy.yaml'].includes(file.basename)
+                )
+            ).toMatchSnapshot();
+        });
+
+        test('enable Fiori tools', async () => {
+            const projectDir = join(outputDir, 'fiori-tools');
+            await generate(
+                projectDir,
+                {
+                    ...config,
+                    deploy: {
+                        package: '$TMP'
+                    },
+                    options: {
+                        fioriTools: true
+                    }
+                },
+                fs
+            );
+            expect(
+                fs.dump(
+                    projectDir,
+                    (file) =>
+                        file.dirname === projectDir &&
+                        ['package.json', 'ui5.yaml', 'ui5-deploy.yaml'].includes(file.basename)
+                )
+            ).toMatchSnapshot();
         });
     });
 });